+++ api/bruno/9Dog-final/02-로그인/로그인.bru
@@ -12,7 +12,7 @@ post {
 
 body:json {
   {
-    "email": "tjwngml@naver.com",
+    "email": "admin@gmail.com",
     "password": "123123"
   }
 }
